// src/components/CardCont/media.type.ts
export interface MediaTypes {
    id: string;
    name: string;
    description: string | null;
    path: string;
    filename: string;
    type: string;
    size: number;
    durantion: number;
    created_at: string;
    updated_at: string;
    start_date: string | null;
    expiry_date: string | null;
  }
  